// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      contractor_availability: {
        Row: {
          created_at: string
          friday: boolean | null
          id: string
          monday: boolean | null
          owner_id: string | null
          saturday: boolean | null
          sunday: boolean | null
          thursday: boolean | null
          tuesday: boolean | null
          updated_at: string
          wednesday: boolean | null
        }
        Insert: {
          created_at?: string
          friday?: boolean | null
          id?: string
          monday?: boolean | null
          owner_id?: string | null
          saturday?: boolean | null
          sunday?: boolean | null
          thursday?: boolean | null
          tuesday?: boolean | null
          updated_at?: string
          wednesday?: boolean | null
        }
        Update: {
          created_at?: string
          friday?: boolean | null
          id?: string
          monday?: boolean | null
          owner_id?: string | null
          saturday?: boolean | null
          sunday?: boolean | null
          thursday?: boolean | null
          tuesday?: boolean | null
          updated_at?: string
          wednesday?: boolean | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}